requirements.txt
Flask
Flask-Caching
Flask-JWT-Extended
Flask-Mail
Flask-SQLAlchemy
Flask-WTF
Werkzeug
psycopg2-binary
requests
flask-migrate
email_validator
dotenv
Pillow
playwright
owslib
pyproj
certifi
aiohttp
urllib3
certifi
gunicorn